refactor: organize sidebar logic into context, provider, and custom hook
