+++ src/main.tsx
@@ -1,13 +1,13 @@
+import { SidebarProvider } from "./providers/SidebarProvider";
 import { BrowserRouter, Route, Routes } from "react-router";
 import { createRoot } from "react-dom/client";
+import NotFound from "./pages/NotFound";
+import Settings from "./pages/Settings";
 import { StrictMode } from "react";
 import Chat from "./pages/Chat";
 import Home from "./pages/Home";
 import Layout from "./layouts";
 import "./styles/global.css";
-import { SidebarProvider } from "./context/SidebarContext";
-import NotFound from "./pages/NotFound";
-import Settings from "./pages/Settings";
 
 export function App() {
   return (
